integrate crawler package with main package to fix gcloud import errors
